testScript issues resolved | Building tests
