+++ coldfire/anchor/package.json
@@ -9,13 +9,8 @@
   "module": "./index.js",
   "private": true,
   "devDependencies": {
-    "@types/chai": "^4.3.19",
-    "chai": "^5.1.1",
-    "ts-mocha": "^10.0.0",
-    "typescript": "^5.0.4",
-    "mocha": "^10.2.0"
-  },
-  "scripts": {
-    "test": "ANCHOR_PROVIDER_URL='http://127.0.0.1:8899' ts-mocha -p ./tsconfig.json -t 1000000 tests/**/*.ts"
+    "@types/jest": "^29.5.12",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.5.4"
   }
 }
