+++ coldfire/anchor/tsconfig.json
@@ -1,8 +1,7 @@
 {
   "extends": "../tsconfig.base.json",
   "compilerOptions": {
-    "module": "commonjs",
-    "types": ["chai", "mocha"]
+    "module": "commonjs"
   },
   "files": [],
   "include": [],
