+++ coldfire/package.json
@@ -49,7 +49,7 @@
     "@swc/helpers": "~0.5.2",
     "@swc/jest": "0.2.20",
     "@types/chai": "^4.3.19",
-    "@types/jest": "^29.4.0",
+    "@types/jest": "^29.5.12",
     "@types/mocha": "^10.0.7",
     "@types/node": "18.16.9",
     "@types/react": "18.3.1",
